[ts-sdk-v2] Move paginate with cursor to client
